% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_input_sequence.R
\name{analyze_input_sequence}
\alias{analyze_input_sequence}
\title{Analyze Input Sequence}
\usage{
analyze_input_sequence(input_sequence, ref_files)
}
\arguments{
\item{input_sequence}{A string representing the input DNA sequence.}

\item{ref_files}{A character vector of file paths to reference DNA FASTA files.}
}
\value{
A data frame with columns:
\item{Reference_Name}{The name of the reference sequence.}
\item{Alignment_Score}{The alignment score for the input sequence.}
}
\description{
This function aligns an input DNA sequence against a set of reference sequences and returns a sorted data frame of alignment scores.
}
\examples{
input_sequence <- "ATCGTACG"
ref_files <- c("file1.fasta", "file2.fasta")
results <- analyze_input_sequence(input_sequence, ref_files)
print(results)
}
