% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_reference_sequences.R
\name{load_reference_sequences}
\alias{load_reference_sequences}
\title{Load Reference Sequences}
\usage{
load_reference_sequences(file_paths)
}
\arguments{
\item{file_paths}{A character vector of file paths to FASTA files containing DNA sequences.}
}
\value{
A \code{DNAStringSet} object containing the loaded DNA sequences with cleaned headers.
}
\description{
This function loads reference DNA sequences from multiple FASTA files and standardizes their names.
}
\examples{
file_paths <- c("file1.fasta", "file2.fasta")
ref_seqs <- load_reference_sequences(file_paths)
print(ref_seqs)
}
