% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_alignment.R
\name{perform_alignment}
\alias{perform_alignment}
\title{Perform Pairwise Alignment}
\usage{
perform_alignment(input_seq, ref_seq)
}
\arguments{
\item{input_seq}{A string representing the input DNA sequence.}

\item{ref_seq}{A string representing the reference DNA sequence.}
}
\value{
A list containing:
\item{score}{The alignment score.}
\item{alignment}{The pairwise alignment object.}
}
\description{
This function performs local pairwise alignment between an input sequence and a reference sequence.
}
\examples{
input_seq <- "ATCG"
ref_seq <- "ATGG"
result <- perform_alignment(input_seq, ref_seq)
print(result$score)
}
